Add benchmarks for many libraries

+++ benchmark/src/utils.h
@@ -1,5 +1,6 @@
 #pragma once
 
 #include <string>
+#include <cstdint>
 
-std::string generate_random_number_string(int digits);
+std::string generate_random_number_string(int64_t digits);
